+++ src/app/admin/components/surveillance/edit.controller.js
@@ -54,7 +54,7 @@
 
         function addRequirement () {
             vm.modalInstance = $uibModal.open({
-                templateUrl: 'app/admin/components/surveillance/editRequirement.html',
+                templateUrl: 'app/admin/components/surveillance/requirement/edit.html',
                 controller: 'EditRequirementController',
                 controllerAs: 'vm',
                 animation: false,
@@ -100,7 +100,7 @@
         function editRequirement (req) {
             req.guiId = req.id ? req.id : (new Date()).getTime();
             vm.modalInstance = $uibModal.open({
-                templateUrl: 'app/admin/components/surveillance/editRequirement.html',
+                templateUrl: 'app/admin/components/surveillance/requirement/edit.html',
                 controller: 'EditRequirementController',
                 controllerAs: 'vm',
                 animation: false,
@@ -142,7 +142,7 @@
 
         function inspectNonconformities (noncons) {
             vm.modalInstance = $uibModal.open({
-                templateUrl: 'app/admin/components/surveillance/nonconformityInspect.html',
+                templateUrl: 'app/admin/components/surveillance/nonconformity/inspect.html',
                 controller: 'NonconformityInspectController',
                 controllerAs: 'vm',
                 animation: false,