+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@ _Date TBD_
 * Ellipsis long names on CMS ID & Compare widgets
 * Links to Listing details from CAP Collection open "Surveillance Activity" accordion
 * Added "Measures Successfully Tested" accordion element
+* Updated CA Collection table to multi-sort
 
 ### Text changes
 * Change "Additional Software" to "Relied Upon Software"