+++ RELEASE_NOTES.md
@@ -1,7 +1,7 @@
 # Release Notes
 
-## Version TBD
-_Date TBD_
+## Version 9.1.0
+_3 July 2017_
 
 ### New features
 * ICS Family Tree work